Refactor web UI into components
This will help maintainability
