+++ src/web/.gitignore
@@ -3,3 +3,4 @@ dist/
 *.log
 .DS_Store
 *.swp
+.vite
